LayoutTests/inspector/console-format-collections-expected.txt

CONSOLE MESSAGE: line 15: [object NodeList]
CONSOLE MESSAGE: line 19: [object HTMLCollection]
CONSOLE MESSAGE: line 23: [object HTMLOptionsCollection]
CONSOLE MESSAGE: line 27: [object HTMLAllCollection]
CONSOLE MESSAGE: line 31: [object HTMLCollection]
CONSOLE MESSAGE: line 35: [object NodeList]
Tests that console nicely formats HTML Collections and NodeLists.

console-format-collections.html:15[<select id="sel" name="sel">]
console-format-collections.html:19[<script src="inspector-test.js">, <script src="console-tests.js">, <script>]
console-format-collections.html:23[<option value="1">one</option>, <option value="2">two</option>]
console-format-collections.html:27[<html>, <head>, <script src="inspector-test.js">, <script src="console-tests.js">, <script>, <body onload="onload()">, <p> Tests that console nicely formats HTML Collections and NodeLists. </p>, <div style="display:none">, <form id="f">, <select id="sel" name="sel">, <option value="1">one</option>, <option value="2">two</option>, <input type="radio" name="x" value="x1">, <input type="radio" name="x" value="x2">, <div id="output">]
console-format-collections.html:31[<select id="sel" name="sel">, <input type="radio" name="x" value="x1">, <input type="radio" name="x" value="x2">]
console-format-collections.html:35[<input type="radio" name="x" value="x1">, <input type="radio" name="x" value="x2">]
